+++ Sources/Swift/Options+Dictionary.swift
@@ -194,10 +194,6 @@ extension Options {
             self.reportAccessibilityIdentifier = reportAccessibilityIdentifier
         }
 
-        if let enableBreadcrumbTextExtraction = boolValue(dictionary["enableBreadcrumbTextExtraction"]) {
-            self.enableBreadcrumbTextExtraction = enableBreadcrumbTextExtraction
-        }
-
         if let enableUserInteractionTracing = boolValue(dictionary["enableUserInteractionTracing"]) {
             self.enableUserInteractionTracing = enableUserInteractionTracing
         }