+++ Tests/SentryTests/Integrations/Breadcrumbs/SentryBreadcrumbTrackingIntegrationTests.swift
@@ -91,10 +91,9 @@ class SentryBreadcrumbTrackingIntegrationTests: XCTestCase {
     }
 
 #if os(iOS)
-    func testInit_whenBreadcrumbTextExtractionDisabled_shouldPassOptionToTracker() throws {
+    func testInit_whenBreadcrumbTextExtractionIsDefault_shouldDisableTracker() throws {
         // -- Arrange --
         let options = fixture.defaultOptions
-        options.enableBreadcrumbTextExtraction = false
 
         // -- Act --
         let sut = try fixture.getSut(options: options)
@@ -107,6 +106,24 @@ class SentryBreadcrumbTrackingIntegrationTests: XCTestCase {
 
         // -- Assert --
         XCTAssertFalse(tracker.enableBreadcrumbTextExtraction)
+    }
+
+    func testInit_whenBreadcrumbTextExtractionIsEnabled_shouldEnableTracker() throws {
+        // -- Arrange --
+        let options = fixture.defaultOptions
+        options.experimental.enableBreadcrumbTextExtraction = true
+
+        // -- Act --
+        let sut = try fixture.getSut(options: options)
+        defer {
+            sut.uninstall()
+        }
+        let tracker = try XCTUnwrap(
+            Mirror(reflecting: sut).descendant("breadcrumbTracker") as? SentryBreadcrumbTracker
+        )
+
+        // -- Assert --
+        XCTAssertTrue(tracker.enableBreadcrumbTextExtraction)
     }
 #endif
 }